requirements.txt
streamlit>=1.30
pymupdf>=1.23.20
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=2.2.2
openai>=1.14.3
tiktoken>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
openpyxl>=3.1.0
beautifulsoup4